slide-generation/requirements.txt
flask
pillow
python-pptx